fix macos build

+++ test/long_arithmetic/unsigned_big_integer_test.cpp
@@ -173,10 +173,10 @@ TYPED_TEST_P(UBigIntegerTest, division_test) {
   EXPECT_EQ(val1 / val1, UBigIntegerT(1));
   EXPECT_EQ(val2 / val2, UBigIntegerT(1));
 
-  val1 = 41346146161;
+  val1 = (uint64_t)41346146161;
   val2 = 10;
 
-  EXPECT_EQ(val1 / val2, UBigIntegerT(4134614616));
+  EXPECT_EQ(val1 / val2, UBigIntegerT((uint64_t)4134614616));
   EXPECT_EQ(val2 / val1, UBigIntegerT(0));
   EXPECT_EQ(val1 / val1, UBigIntegerT(1));
   EXPECT_EQ(val2 / val2, UBigIntegerT(1));
@@ -243,7 +243,7 @@ TYPED_TEST_P(UBigIntegerTest, division_test2) {
   EXPECT_EQ(val1 % val1, UBigIntegerT(0));
   EXPECT_EQ(val2 % val2, UBigIntegerT(0));
 
-  val1 = 41346146161;
+  val1 = (uint64_t)41346146161;
   val2 = 10;
 
   EXPECT_EQ(val1 % val2, UBigIntegerT(1));
